Define SZ_MAX consistently with pecos
This change has already been made on the json_parser branch
of the main Dakota repo. When the json_parser branch of the
pecos repo was merged to devel, it became necessary to do it
here, too.

+++ src/dakota_global_defs.hpp
@@ -21,7 +21,7 @@ namespace Dakota {
 // --------------
 /// special value returned by index() when entry not found
 //const size_t _NPOS  = ~(size_t)0; // one's complement
-const size_t SZ_MAX = std::numeric_limits<size_t>::max();
+const size_t SZ_MAX = std::numeric_limits<int64_t>::max();
 const size_t _NPOS  = SZ_MAX; // alias old definition
 /// bound beyond which constraints are considered inactive
 const double BIG_REAL_BOUND = 1.0e+30;
